+++ client/src/components/home/dashboard/Dashboard.js
@@ -20,7 +20,7 @@ import MenuIcon from "@material-ui/icons/Menu";
 import ChevronLeftIcon from "@material-ui/icons/ChevronLeft";
 import ExitToAppIcon from "@material-ui/icons/ExitToApp";
 import { mainListItems, secondaryListItems } from "./listItems";
-import Chart from "./Chart";
+import Chart from "../../expenses/Chart";
 import Income from "../../income/Income";
 import ExpenseTable from "../../expenses/ExpenseTable";
